editor/libeditor/text/tests/test_bug590554.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=590554
-->

<head>
  <title>Test for Bug 590554</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>  
</head>

<body>

  <script type="application/javascript">

    /** Test for Bug 590554 **/

    SimpleTest.waitForExplicitFinish();

    SimpleTest.waitForFocus(function() {
      var t = document.querySelector("textarea");
      t.focus();
      synthesizeKey("VK_ENTER", {});
      is(t.value, "\n", "Pressing enter should work the first time");
      synthesizeKey("VK_ENTER", {});
      is(t.value, "\n", "Pressing enter should not work the second time");
      SimpleTest.finish();
    });

  </script>

  <textarea maxlength="1"></textarea>
</body>
</html>
